+++ src/Metadata/PropertyType.php
@@ -19,6 +19,11 @@ interface PropertyType extends \Stringable
     public function isNullable(): bool;
 
     /**
+     * Creates a new instance with the given nullability
+     */
+    public function asNullable(bool $nullable = true): self;
+
+    /**
      * Merges another property type into this one.
      *
      * @param PropertyType<*> $other